+++ test/test_entity.py
@@ -1,6 +1,6 @@
 import unittest
 
-from iolite_client.entity import Heating, Room, Switch
+from iolite_client.entity import Heating, RadiatorValve, Room, Switch
 
 
 class RoomTest(unittest.TestCase):
@@ -8,6 +8,16 @@ class RoomTest(unittest.TestCase):
         self.bedroom = Room("placeIdentifier-1", "Bedroom")
         self.bedroom_switch = Switch(
             "2", "Bedroom Switch", self.bedroom.identifier, "Generic"
+        )
+        self.bedroom_radiator_valve = RadiatorValve(
+            "3",
+            "Bedroom Radiator Valve",
+            self.bedroom.identifier,
+            "Manu",
+            21.5,
+            55,
+            "Normal",
+            0.0,
         )
         self.bedroom_heating = Heating(self.bedroom.identifier, "Bedroom", 20.0, 21.5)
 
@@ -19,6 +29,25 @@ class RoomTest(unittest.TestCase):
         self.assertEqual(1, len(self.bedroom.devices))
         self.assertEqual(
             self.bedroom_switch, self.bedroom.devices[self.bedroom_switch.identifier]
+        )
+
+    def test_add_device_again(self):
+        self.bedroom.add_device(self.bedroom_radiator_valve)
+        bedroom_radiator_valve = self.bedroom.devices[
+            self.bedroom_radiator_valve.identifier
+        ]
+        self.assertEqual(
+            bedroom_radiator_valve.current_env_temp,
+            21.5,
+        )
+        bedroom_radiator_valve.current_env_temp = 25
+        self.bedroom.add_device(bedroom_radiator_valve)
+        bedroom_radiator_valve = self.bedroom.devices[
+            self.bedroom_radiator_valve.identifier
+        ]
+        self.assertEqual(
+            bedroom_radiator_valve.current_env_temp,
+            25,
         )
 
     def test_room_has_device(self):