+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+
+- Fix type on `valve_position` on `RadiatorValve`
+
 ## 0.2.0 - 2021-10-07
 
 ### Added